Апгрейд

+++ Sortings.h
@@ -5,8 +5,8 @@
 #include <stdio.h>
 
 void swap_lineinfo(struct LineInfo *arr1, struct LineInfo *arr2);
-size_t partition(struct LineInfo *text_ptr, size_t l, size_t r, int (*compares)(const void *, const void *));
-void quick_sort(struct LineInfo *text_ptr, size_t l, size_t r, int (*compares)(const void *, const void *));
+int partition(struct LineInfo *text_ptr, int l, int r, int (*compares)(const void *, const void *));
+void quick_sort(struct LineInfo *text_ptr, int l, int r, int (*compares)(const void *, const void *));
 
 void bubble_sort (LineInfo *text_ptr);
 int Mstrcmp_LtoR(char *str1, char *str2, char *end_ptr1, char *end_ptr2);
